Move localization.dart from util to l10n
